+++ components/portal/misc/directives.js
@@ -239,33 +239,6 @@ define(['angular', 'require'], function(angular, require) {
       };
     })
 
-  /**
-   * Launch Button Directive
-   * Displays a button that fits the width and visual style of a widget
-   * Template:
-   *  <launch-button data-href="" data-target=""
-   *    data-button-text="" data-aria-label=""></launch-button>
-   *
-   * Params:
-   * - href: where you want them to go
-   * - target: open in new window, new tab, or same window
-   * - button-text: the text to be displayed
-   * - aria-label: (optional) text to provide additional
-   *    context for screen readers, if necessary
-   */
-    .directive('launchButton', function() {
-      return {
-        restrict: 'E',
-        scope: {
-          href: '@href',
-          target: '@target',
-          buttonText: '@buttonText',
-          ariaLabel: '@ariaLabel',
-        },
-        templateUrl: require.toUrl('./partials/launch-button.html'),
-      };
-    })
-
     .directive('addToHome', function() {
       return {
         restrict: 'E',